Added a feature to reverse sort list of names
Added an if statement to reverse the order of the sorted list. Output based on menu choice of user's desired sort order. Menu has a do-while loop to continue running until a valid menu option is selected and a try-catch to ensure that only integer values are accepted for input and that the program can continue working if an integer value is inputted.
Also updated Instructions.txt with directions on how to use new reverse sort option.